+++ Nexus/src/ECS/Component.h
@@ -3,7 +3,7 @@
 struct IComponent
 {
 protected:
-	static int m_nextId;
+	static size_t m_nextId;
 };
 
 template <typename T>
@@ -11,7 +11,7 @@ class Component : public IComponent
 {
 public:
 	/* The first time GetId() is called for a Component<T>, it'll create a unique id which will remain same for each subsequent calls since id is also static */
-	static int GetId()
+	static size_t GetId()
 	{
 		static auto id = m_nextId++;
 		return id;
